instances __the_amount_of_sugar_in_the_juice_per_serving and juice were NOT merged together
Query: in the Juice class, should the instances '- the amount of sugar in the juice per serving' and 'Juice' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No. While both concepts may be related to each other (with one describing an attribute of the other), they still refer to distinct entities and thus are not semantic or ontological duplicates. Merging them would result in conflating two different ideas and could lead to confusion.
meaning that instances __the_amount_of_sugar_in_the_juice_per_serving and juice are different.
Query cache in files:
- /kgfiller/data/cache-110cef615d6d86690e87f567d7e8fe5dcf25f0f766573293cfd2b701b7b73919.yml