instances __the_amount_of_protein_in_the_juice_per_serving and tomato_juice_as_needed were NOT merged together
Query: in the Juice class, should the instances '- the amount of protein in the juice per serving' and 'tomato juice as needed' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

The amount of protein in the juice per serving is a specific piece of information about the nutritional content of the juice, while "tomato juice as needed" refers to the ingredient used to make the juice. These two pieces of information serve different purposes and are not duplicative in terms of their meaning (semantics) or their relationship to other concepts (ontology).
meaning that instances __the_amount_of_protein_in_the_juice_per_serving and tomato_juice_as_needed are different.
Query cache in files:
- /kgfiller/data/cache-71074efbcdf149ba9fca1cc7211f55aec95b10153cdb6920c60b479aef799a29.yml